+++ .solcover.js
@@ -17,6 +17,7 @@ module.exports = {
     },
     skipFiles: [
         'tools/Migrations.sol',
+        'tools/testing/ERC1271Mock.sol',
         'tools/testing/TestClient.sol',
         'tools/testing/TestReceiver.sol',
         'modules/delegates/SignatureVerifier.sol',